refactored directory names
