--- src/integrationTest/java/uk/gov/hmcts/reform/adoption/ApplicationTest.java
@@ -1,14 +1,0 @@
-package uk.gov.hmcts.reform.adoption;
-
-import org.junit.jupiter.api.Test;
-
-public class ApplicationTest {
-    @Test
-    public void applicationContextLoaded() {
-    }
-
-    @Test
-    public void applicationContextTest() {
-        Application.main(new String[] {});
-    }
-}
