+++ src/integrationTest/java/uk/gov/hmcts/reform/adoption/controller/GetWelcomeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.adoption.controllers;
+package uk.gov.hmcts.reform.adoption.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
