+++ src/main/java/uk/gov/hmcts/reform/adoption/controller/StartCaseController.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.adoption.controllers;
+package uk.gov.hmcts.reform.adoption.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.reform.adoption.services.CoreCaseDataService;
+import uk.gov.hmcts.reform.adoption.service.CoreCaseDataService;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 
 @RestController
