+++ src/main/java/uk/gov/hmcts/reform/adoption/model/Applicant.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.adoption.models;
+package uk.gov.hmcts.reform.adoption.model;
 
 import lombok.Builder;
 import lombok.Data;
